requirements.txt
git+https://github.com/m-bain/whisperx.git
numpy<2
python-dotenv==1.0.1